fixed compilation warnings

+++ qmail-1.03/substdo.c
@@ -17,7 +17,7 @@
 
 static int
 allwrite(op, fd, buf, len)
-	register int    (*op) ();
+	register ssize_t (*op) ();
 	register int    fd;
 	register char  *buf;
 	register int    len;
